backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
PyGithub==2.4.0
python-dotenv==1.0.1
pydantic==2.9.2
gitpython==3.1.43
aiosqlite==0.20.0
